# IAMApp/IAM/Models/UserName.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace IAM.Models
{
    public class RegisterUser
    {
        public string Email { get; set; }

        public string Password { get; set; }

        //public string ConfirmPassword { get; set; }



    }
}